stuff till transformation

+++ ACCAT_OGL/Shader.h
@@ -13,6 +13,6 @@ public:
 	void use();
 	void set_bool(const std::string& name, bool value) const;
 	void set_int(const std::string& name, int value) const;
-	void set_float(const std::string& name, int value) const;
+	void set_float(const std::string& name, float value) const;
 	void set_4f(const std::string& name, float f0, float f1, float f2, float f3) const;
 };
